server/mcp_server_tag/pyproject.toml
[project]
name = "mcp-server-tag"
version = "1.0.0"
description = "A Model Context Protocol server providing tools for volcengin tag for LLMs"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "volc-sdk-team", email = "volc-sdk-team@bytedance.com" },
]
keywords = ["tag", "mcp", "llm"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "aiohttp>=3.11.18",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tzdata>=2024.2",
    "volcengine-python-sdk>=3.0.1",
]

[project.scripts]
mcp-server-tag = "mcp_server_tag:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "freezegun>=1.5.1",
    "pyright>=1.1.389",
    "pytest>=8.3.3",
    "ruff>=0.8.1",
]
